only delete cmakecache on clean

+++ automaton_2d/build.sh
@@ -6,13 +6,24 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 # Navigate to the project directory
 cd $DIR
 
-# Remove the build directory if it exists
-if [ -d "build" ]; then
-    rm -r build
+# If the argument is "clean", remove the entire build directory
+if [ "$1" == "clean" ]; then
+    if [ -d "build" ]; then
+        rm -r build
+    fi
+    exit 0
 fi
 
-# Create a new build directory and navigate into it
-mkdir build
+# If the build directory doesn't exist, create it
+if [ ! -d "build" ]; then
+    mkdir build
+fi
+
+# If CMakeCache.txt exists, empty it to force reconfiguration
+if [ -f "build/CMakeCache.txt" ]; then
+    > build/CMakeCache.txt
+fi
+
 cd build
 
 # Check for RtAudio argument
@@ -22,7 +33,6 @@ if [ "$1" ]; then
 else
     cmake ..
 fi
-
 
 # If CMake completes without errors, build the project
 if [ $? -eq 0 ]; then
